removing FastCGI support

+++ components/tools/OmeroPy/src/omero/plugins/web.py
@@ -15,6 +15,7 @@ import platform
 import sys
 import os
 import re
+from functools import wraps
 from omero_ext.argparse import SUPPRESS
 
 HELP = "OMERO.web configuration/deployment tools"
@@ -32,9 +33,10 @@ Configuration:
 Example Nginx developer usage:
 
     omero config set omero.web.debug true
+    omero config set omero.web.application_server.max_requests 1
     omero web config nginx-development --http=8000 >> nginx.conf
-    omero web start
     nginx -c `pwd`/nginx.conf
+    omero web start --wsgi-args ' --reload'
     omero web status
     omero web stop
     nginx -s stop
@@ -53,11 +55,64 @@ Example IIS usage:
 """
 
 
+def config_required(func):
+    """Decorator validating Django dependences and omeroweb/settings.py"""
+    def import_django_settings(func):
+        def wrapper(self, *args, **kwargs):
+            try:
+                import django  # NOQA
+            except:
+                self.ctx.die(681, "Django not installed!")
+            try:
+                import omeroweb.settings as settings
+                kwargs['settings'] = settings
+            except Exception, e:
+                self.ctx.die(682, e)
+            return func(self, *args, **kwargs)
+        return wrapper
+    return wraps(func)(import_django_settings(func))
+
+
+def assert_config_argtype(func):
+    """Decorator validating OMERO.web deployment dependences"""
+    def config_argtype(func):
+        def wrapper(self, *args, **kwargs):
+            argtype = args[0].type
+            settings = kwargs['settings']
+            mismatch = False
+            if args[0].system:
+                self.ctx.die(683,
+                             "ERROR: --system is no longer supported, "
+                             "see --help")
+            if settings.APPLICATION_SERVER in ("development",):
+                mismatch = True
+            if settings.APPLICATION_SERVER in (settings.WSGITCP,):
+                try:
+                    import gunicorn  # NOQA
+                except ImportError:
+                    self.ctx.die(690,
+                                 "ERROR: FastCGI support was removed in "
+                                 "OMERO 5.2. Install Gunicorn and update "
+                                 "config.")
+                if argtype not in ("nginx", "nginx-development",):
+                    mismatch = True
+            if (settings.APPLICATION_SERVER in (settings.WSGI,) and
+                    argtype not in ("apache",)):
+                mismatch = True
+            if mismatch:
+                self.ctx.die(680,
+                             ("ERROR: configuration mismatch. "
+                              "omero.web.application_server=%s cannot be "
+                              "used with 'omero web config %s'.") %
+                             (settings.APPLICATION_SERVER, argtype))
+            return func(self, *args, **kwargs)
+        return wrapper
+    return wraps(func)(config_argtype(func))
+
+
 class WebControl(BaseControl):
 
-    config_choices = ("nginx", "nginx-development",
-                      "nginx-wsgi", "nginx-wsgi-development",
-                      "apache", "apache-fcgi", "apache-wsgi")
+    config_choices = ("nginx", "nginx-development", "apache")
 
     def _configure(self, parser):
         sub = parser.sub()
@@ -101,11 +156,7 @@ class WebControl(BaseControl):
             "Output a config template for web server\n"
             "  nginx: Nginx system configuration for inclusion\n"
             "  nginx-development: Standalone user-run Nginx server\n"
-            "  nginx-wsgi: Standalone user-run Nginx server\n"
-            "  nginx-wsgi-development: Standalone user-run Nginx server\n"
-            "  apache: Apache 2.2 with mod_fastcgi\n"
-            "  apache-fcgi: Apache 2.4+ with mod_proxy_fcgi\n"
-            "  apache-wsgi: Apache 2.2+ with mod_wsgi\n")
+            "  apache: Apache 2.2+ with mod_wsgi\n")
         config.add_argument("type", choices=self.config_choices)
         nginx_group = config.add_argument_group(
             'Nginx arguments', 'Optional arguments for nginx templates.')
@@ -155,9 +206,9 @@ class WebControl(BaseControl):
             "Developer use: Loads the blitz gateway into a Python"
             " interpreter")
 
-    def help(self, args):
+    @config_required
+    def help(self, args, settings):
         """Return extended help"""
-        from omeroweb import settings
         try:
             CONFIG_TABLE_FMT = "    %-35.35s  %-8s  %r\n"
             CONFIG_TABLE = CONFIG_TABLE_FMT % (
@@ -176,39 +227,14 @@ class WebControl(BaseControl):
 
         self.ctx.err(LONGHELP % CONFIG_TABLE)
 
-    def _check_lib_installed(self):
-        try:
-            import django  # NOQA
-        except:
-            self.ctx.err("Django not installed!")
-
     def _get_python_dir(self):
         return self.ctx.dir / "lib" / "python"
 
     def _get_fallback_dir(self):
         return self.ctx.dir / "lib" / "fallback"
 
-    def _get_templates_dir(self):
-        return self.ctx.dir / "etc" / "templates"
-
-    def _set_nginx_fastcgi(self, d, settings):
-        script_info = (
-            "fastcgi_split_path_info ^(%s)(.*)$;\n"
-            "            fastcgi_param PATH_INFO $fastcgi_path_info;\n"
-            "            fastcgi_param SCRIPT_INFO $fastcgi_script_name;\n")
-        script_info_fallback = (
-            "fastcgi_param PATH_INFO $fastcgi_script_name;\n")
-        try:
-            d["FASTCGI_PATH_SCRIPT_INFO"] = (
-                script_info % settings.FORCE_SCRIPT_NAME)
-        except:
-            d["FASTCGI_PATH_SCRIPT_INFO"] = script_info_fallback
-
-    def _set_apache_fcgi_fastcgi(self, d, settings):
-        # OMERO.web requires the fastcgi PATH_INFO variable, which
-        # mod_proxy_fcgi obtains by taking everything after the last
-        # path component containing a dot.
-        d["CGI_PREFIX"] = "%s.fcgi" % d["FORCE_SCRIPT_NAME"]
+    def _get_web_templates_dir(self):
+        return self.ctx.dir / "etc" / "templates" / "web"
 
     def _set_apache_wsgi(self, d, settings):
         # WSGIDaemonProcess requires python-path and user
@@ -228,54 +254,22 @@ class WebControl(BaseControl):
         d["OMEROPYTHONROOT"] = self._get_python_dir()
         d["OMEROFALLBACKROOT"] = self._get_fallback_dir()
 
-    def _fastcgi_deprecation(self, settings):
-        if settings.APPLICATION_SERVER in settings.FASTCGI_TYPES:
-            self.ctx.err(
-                "WARNING: FastCGI support is deprecated and will be removed"
-                " in OMERO 5.2. Install Gunicorn and update config.")
-
-    def _assert_config_argtype(self, argtype, settings):
-        mismatch = False
-        if (settings.APPLICATION_SERVER in settings.WSGI_TYPES and
-                "wsgi" not in argtype):
-            mismatch = True
-        if settings.APPLICATION_SERVER in ("development",):
-            mismatch = True
-        if (settings.APPLICATION_SERVER in settings.FASTCGI_TYPES and
-                argtype not in ("nginx", "nginx-development",
-                                "apache", "apache-fcgi")):
-            mismatch = True
-        if mismatch:
-            self.ctx.die(680, ("ERROR: configuration mismatch. "
-                               "omero.web.application_server=%s cannot be"
-                               " used with 'omero web config %s'"
-                               ".") % (settings.APPLICATION_SERVER, argtype))
-
-    def config(self, args):
+    @config_required
+    @assert_config_argtype
+    def config(self, args, settings):
         """Generate a configuration file from a template"""
-        from omeroweb import settings
-
-        if args.system:
-            self.ctx.err(
-                "WARNING: --system is no longer supported, see --help")
-
-        self._assert_config_argtype(args.type, settings)
-
         server = args.type
         if args.http:
             port = args.http
-        elif server in ('nginx-development', 'nginx-wsgi-development'):
+        elif server in ('nginx-development',):
             port = 8080
         else:
             port = 80
 
-        self._fastcgi_deprecation(settings)  # to be removed in 5.2
-
-        if settings.APPLICATION_SERVER in (settings.FASTCGITCP,
-                                           settings.WSGITCP):
+        if settings.APPLICATION_SERVER in settings.WSGITCP:
             if settings.APPLICATION_SERVER_PORT == port:
                 self.ctx.die(
-                    678, "Port conflict: HTTP(%s) and"" wsgi/fastcgi-tcp(%s)."
+                    678, "Port conflict: HTTP(%s) and"" wsgi(%s)."
                     % (port, settings.APPLICATION_SERVER_PORT))
 
         d = {
@@ -285,12 +279,10 @@ class WebControl(BaseControl):
             "NOW": str(datetime.now())}
 
         if server in ("nginx", "nginx-development",
-                      "nginx-wsgi", "nginx-wsgi-development",
-                      "apache-wsgi"):
+                      "apache",):
             d["HTTPPORT"] = port
 
-        if server in ("nginx", "nginx-development",
-                      "nginx-wsgi", "nginx-wsgi-development"):
+        if server in ("nginx", "nginx-development",):
             d["MAX_BODY_SIZE"] = args.max_body_size
 
         # FORCE_SCRIPT_NAME always has a starting /, and will not have a
@@ -305,45 +297,34 @@ class WebControl(BaseControl):
             d["FORCE_SCRIPT_NAME"] = "/"
             d["PREFIX_NAME"] = ""
 
-        if server in ("apache", "apache-fcgi", "apache-wsgi"):
+        if server in ("apache",):
             try:
                 d["WEB_PREFIX"] = settings.FORCE_SCRIPT_NAME.rstrip("/")
             except:
-                if server == "apache-wsgi":
-                    d["WEB_PREFIX"] = "/"
-                else:
-                    d["WEB_PREFIX"] = ""
-
-        if settings.APPLICATION_SERVER not in (settings.FASTCGI_TYPES +
-                                               settings.WSGI_TYPES):
-            self.ctx.die(
-                679,
-                "Web template configuration requires"
-                " fastcgi-tcp or wsgi/wsgi-tcp.")
+                d["WEB_PREFIX"] = "/"
 
         d["FASTCGI_EXTERNAL"] = '%s:%s' % (
             settings.APPLICATION_SERVER_HOST, settings.APPLICATION_SERVER_PORT)
 
-        if server in ("nginx", "nginx-development"):
-            self._set_nginx_fastcgi(d, settings)
+        if settings.APPLICATION_SERVER not in settings.WSGI_TYPES:
+            self.ctx.die(679,
+                         "Web template configuration requires"
+                         "wsgi or wsgi-tcp.")
 
-        if server == "apache-fcgi":
-            self._set_apache_fcgi_fastcgi(d, settings)
-
-        if server == "apache-wsgi":
+        if server == "apache":
             self._set_apache_wsgi(d, settings)
 
         template_file = "%s.conf.template" % server
-        c = file(self._get_templates_dir() / template_file).read()
+        c = file(self._get_web_templates_dir() / template_file).read()
         self.ctx.out(c % d)
 
     def syncmedia(self, args):
         self.collectstatic()
 
-    def enableapp(self, args):
+    @config_required
+    def enableapp(self, args, settings):
         location = self._get_python_dir() / "omeroweb"
         if not args.appname:
-            from omeroweb import settings
             apps = [x.name for x in filter(
                 lambda x: x.isdir() and
                 (x / 'scripts' / 'enable.py').exists(),
@@ -388,26 +369,25 @@ class WebControl(BaseControl):
                 scriptname = scriptname[0]
             cargs.extend([location / appname / "scripts" / scriptname] +
                          args.arg)
-            print cargs
             os.environ['DJANGO_SETTINGS_MODULE'] = 'omeroweb.settings'
             self.set_environ()
             self.ctx.call(cargs, cwd=location)
         except:
             print traceback.print_exc()
 
-    def collectstatic(self):
+    @config_required
+    def collectstatic(self, settings):
         """Ensure that static media is copied to the correct location"""
-        self._check_lib_installed()
         location = self._get_python_dir() / "omeroweb"
         args = [sys.executable, "manage.py", "collectstatic", "--noinput"]
         rv = self.ctx.call(args, cwd=location)
         if rv != 0:
             self.ctx.die(607, "Failed to collect static content.\n")
 
-    def clearsessions(self, args):
+    @config_required
+    def clearsessions(self, args, settings):
         """Clean out expired sessions."""
         self.ctx.out("Clearing expired sessions. This may take some time... ")
-        self._check_lib_installed()
         location = self._get_python_dir() / "omeroweb"
         cmd = [sys.executable, "manage.py", "clearsessions"]
         if not args.no_wait:
@@ -418,21 +398,20 @@ class WebControl(BaseControl):
         else:
             self.ctx.popen(cmd, cwd=location)
 
-    def start(self, args):
+    @config_required
+    def start(self, args, settings):
         self.collectstatic()
         if not args.keep_sessions:
             self.clearsessions(args)
-        import omeroweb.settings as settings
-        self._fastcgi_deprecation(settings)  # to be removed in 5.2
 
         link = ("%s:%d" % (settings.APPLICATION_SERVER_HOST,
                            settings.APPLICATION_SERVER_PORT))
         location = self._get_python_dir() / "omeroweb"
         deploy = getattr(settings, 'APPLICATION_SERVER')
 
-        if deploy == settings.WSGI:
+        if deploy in (settings.WSGI,):
             self.ctx.out("You are deploying OMERO.web using apache and"
-                         " mod_wsgi.")
+                         " mod_wsgi. Add config to Apache and reload.")
             return 1
         else:
             self.ctx.out("Starting OMERO.web... ", newline=False)
@@ -448,12 +427,7 @@ class WebControl(BaseControl):
                 return 1
 
         # 3216
-        if deploy in (settings.FASTCGI_TYPES + (settings.WSGI,)):
-            if "Windows" == platform.system():
-                self.ctx.out("""
-WARNING: Unless you **really** know what you are doing you should NOT be
-using bin\omero web start on Windows with FastCGI.
-""")
+        if deploy in (settings.WSGI,):
             pid_path = self.ctx.dir / "var" / "django.pid"
             pid_num = None
 
@@ -475,23 +449,7 @@ using bin\omero web start on Windows with FastCGI.
                     pid_path.remove()
                     self.ctx.err("Removed stale %s" % pid_path)
 
-        if deploy == settings.FASTCGITCP:
-            cmd = "python manage.py runfcgi workdir=./"
-            cmd += " method=prefork host=%(host)s port=%(port)d"
-            cmd += " pidfile=%(base)s/var/django.pid daemonize=true"
-            cmd += " maxchildren=5 minspare=1 maxspare=5"
-            cmd += " maxrequests=%(maxrequests)d"
-            django = (cmd % {
-                'maxrequests': settings.APPLICATION_SERVER_MAX_REQUESTS,
-                'base': self.ctx.dir,
-                'host': settings.APPLICATION_SERVER_HOST,
-                'port': settings.APPLICATION_SERVER_PORT}).split()
-            rv = self.ctx.popen(args=django, cwd=location)  # popen
-        elif deploy == settings.WSGITCP:
-            try:
-                import gunicorn  # NOQA
-            except ImportError:
-                self.ctx.die(608, "Gunicorn not installed.")
+        if deploy == settings.WSGITCP:
             try:
                 os.environ['SCRIPT_NAME'] = settings.FORCE_SCRIPT_NAME
             except:
@@ -519,11 +477,9 @@ using bin\omero web start on Windows with FastCGI.
         self.ctx.out("[OK]")
         return rv
 
-    def status(self, args):
+    @config_required
+    def status(self, args, settings):
         self.ctx.out("OMERO.web status... ", newline=False)
-        self._check_lib_installed()
-        import omeroweb.settings as settings
-        self._fastcgi_deprecation(settings)  # to be removed in 5.2
 
         deploy = getattr(settings, 'APPLICATION_SERVER')
         cache_backend = getattr(settings, 'CACHE_BACKEND', None)
@@ -532,7 +488,7 @@ using bin\omero web start on Windows with FastCGI.
         else:
             cache_backend = ''
         rv = 0
-        if deploy in (settings.FASTCGI_TYPES + settings.WSGI_TYPES):
+        if deploy in settings.WSGI_TYPES:
             try:
                 f = open(self.ctx.dir / "var" / "django.pid", 'r')
                 pid = int(f.read())
@@ -546,21 +502,20 @@ using bin\omero web start on Windows with FastCGI.
             except:
                 self.ctx.out("[NOT STARTED]")
                 return rv
+        elif deploy in settings.DEVELOPMENT:
+            self.ctx.err(
+                "DEVELOPMENT: You will have to kill processes by hand!")
         else:
-            self.ctx.err("DEVELOPMENT: You will have to check status by hand!")
+            self.ctx.err(
+                "Invalid APPLICATION_SERVER "
+                "(omero.web.application_server = '%s')!" % deploy)
         return rv
 
-    def stop(self, args):
+    @config_required
+    def stop(self, args, settings):
         self.ctx.out("Stopping OMERO.web... ", newline=False)
-        self._check_lib_installed()
-        import omeroweb.settings as settings
         deploy = getattr(settings, 'APPLICATION_SERVER')
-        if deploy in (settings.FASTCGI_TYPES + settings.WSGI_TYPES):
-            if "Windows" == platform.system():
-                self.ctx.out("""
-WARNING: Unless you **really** know what you are doing you should NOT be
-using bin\omero web start on Windows with FastCGI.
-""")
+        if deploy in settings.WSGI_TYPES:
             pid = 'Unknown'
             pid_path = self.ctx.dir / "var" / "django.pid"
             pid_text = "Unknown"
@@ -585,9 +540,14 @@ using bin\omero web start on Windows with FastCGI.
             finally:
                 if pid_path.exists():
                     pid_path.remove()
-        else:
+        elif deploy in settings.DEVELOPMENT:
             self.ctx.err(
                 "DEVELOPMENT: You will have to kill processes by hand!")
+            return False
+        else:
+            self.ctx.err(
+                "Invalid APPLICATION_SERVER "
+                "(omero.web.application_server = '%s')!" % deploy)
             return False
 
     def restart(self, args):
@@ -602,6 +562,7 @@ using bin\omero web start on Windows with FastCGI.
         os.environ['PATH'] = str(os.environ.get('PATH', '.') + ':' +
                                  self.ctx.dir / 'bin')
 
+    @config_required
     def iis(self, args):
         if not (self._isWindows() or self.ctx.isdebug):
             self.ctx.die(2, "'iis' command is for Windows only")
